Helpers for recursive form generation
com.gfredericks.test.chuck.generators/bounded-recursive-gen forms the
starting point for effective generation of nested Hiccup elements that
conform to the malli schema defined in site.fabrcate.prototype.html.

While I initially thought I might have to implement custom subsequences
using html/->hiccup-schema, starting with non-recursive elements and
building on those is more than "good enough" for gen/such-that to
reliably and quickly produce elements that conform to the schema.

This can be used in tests of Hiccup/HTML manipulation code to ensure
that any input to a function intended to manipulate HTML data structures
has output that still conforms to the spec.