Passing all existing tests for paragraph detection
That may be a rather large cond block, but it's far easier to understand
than a deeply nested loop/recur statement.

The recursive calls to parse-paragraphs within its own body are
non-idiomatic and will probably blow up on very large inputs. This is
somewhat mitigated by the fact that the upper bound on the size of
recursive calls is input size + 1, with most of the recursive calls occurring on subsets of the input.

trampoline might be the right function to implement parse-paragraphs
properly and eliminate those recursive calls.