A drop-in replacement for the previous parse fn
It has the same signature and passes all the same tests. Genuinely
suprised at how quickly I was able to achieve feature parity with the
old method. This is basically ready to merge as is, but I may want to
spend more time fleshing out the grammar to make parsing as effective as
possible (the current way seems kinda hacky).