Initial generative testing for pathological template inputs
The generator is pretty rudimentary. A more meaningful refinement would
be to identify a property that can hold for inputs that successfully
parse and therefore check the validity of arbitrary strings instead of
those deliberately constructed to be incorrect.

However, the nice thing about using a malli seqex schema to define the
shape of the inputs is that it can be extended over time.