Fix agent failure/nil error by rebinding state var
This isn't an optimal solution; the filewatcher should be able to pass
in an arbitrary application state agent, not just the one defined in the
write namespace.

There is one remaining issue (that doesn't cause errors): the initial
write in the filewatcher test appears to happen more than once, which
seems at odds with the queue semantics that send is supposed to have.
It's not clear if this is related to the problems caused by nested agent state.