Specify top-level agent as holder of state for rerender
This ensures that a custom agent is always the one used when draft! is
invoked, meaning the test suite of state management behavior will work
without rebinding write/state to the custom test-state var.

Incidentally, testing this revealed insight into the habitual and
difficult-to-reproduce bug with a java.lang.NullPointerException of
'Cannot invoke "String.length()" because "suffix" is null'. When *agent*
was bound to the "inner" context of the file watcher, rerender would
attempt to look within the file watcher agent for the global
configuration containing :site.fabricate.file/template-suffix and raise
the NPE when it found nil.

Changing the binding within draft! such that the rerender function would
modify the *agent* that draft! gets sent to - e.g. the top-level
application state agent - prevented the NPE from re-occurring. This
means that any similar problems coming from asynchronous agent-based
execution may have similar scoping problems, and the way to detect them
is probably just to print *agent* when and where they're detected.