Rework FSM state definitions + log with u/trace
For clarity's sake, it is important to distinguish the labels applied to
the states taken by values in FSMs from the labels applied to the FSMs
themselves. One value may take on different states in different FSMs.
This commit tries to refine the terminology a bit to capture this distinction.

The tests also indicate that, in a manner similar to the way mulog records event
names, a good convention to use for FSM state names is namespaced keywords
rather than strings. This helps ensure uniqueness and retains provenance
information about the state definitions.

Using u/trace within the try block ensures that mulog is aware of any errors
within the execution of the function called when the FSM advances.

Finally, the body of the advance function has been cleaned up a bit to avoid as
many nested if expressions, which makes the flow of the function easier to
understand.