Initial implementation of produce! multimethod
Adding the :site.fabricate.file/output-file key to the entry map only on
a successful write is a simple and effective way of distinguishing success and
failure.

The dictionary definition multimethod metadata may need to be removed to a
secondary metadata path (see
https://danielgregoire.dev/posts/2021-10-15-clojure-src-test-meta/), but it's
worth recording here for the time being, as I think the dictionary definitions
have helped clarify the concepts in my mind, and thus might be useful to users
of Fabricate (once the documentation has been regenerated using this metadata).

As expected, the implementation is straightforward when the steps are clearly delineated