Rewrite tests using matcher-combinators + add fallback parsing
By default, "rewrite-clj does not do anything sophisticated with reader
conditionals"; it simply produces a wrapped (read-string "...") form, which will
fail on reader conditionals by default because it does not pass the necessary
options to read-string. See discussion on GitHub here:
https://github.com/clj-commons/rewrite-clj/issues/67

Edamame is already on the classpath for Fabricate as a transitive dependency of
Malli, so I used to to parse expressions that fail to read. This fallback code
could be replaced with a different library or cleaned up with a better error
message (or both).